moved wavcal.py to spec
